+++ spec/lib/karafka/web/tracking/consumers/listeners/processing_spec.rb
@@ -24,6 +24,7 @@ RSpec.describe_current do
     before do
       listener.on_consumer_consume(event)
       listener.on_consumer_revoke(event)
+      listener.on_consumer_eof(event)
       listener.on_consumer_shutting_down(event)
     end
 
@@ -108,6 +109,25 @@ RSpec.describe_current do
         expect(sampler.jobs).to include("#{caller.id}-shutdown")
       end
     end
+
+    context 'when type is consumer.eofed.error' do
+      let(:type) { 'consumer.eofed.error' }
+
+      it 'expect not to remove the running job from tracked jobs' do
+        listener.on_error_occurred(event)
+        expect(sampler.jobs).to include("#{caller.id}-consume")
+      end
+
+      it 'expect not to remove revoke job for same consumer' do
+        listener.on_error_occurred(event)
+        expect(sampler.jobs).to include("#{caller.id}-revoked")
+      end
+
+      it 'expect not to remove shutdown job for same consumer' do
+        listener.on_error_occurred(event)
+        expect(sampler.jobs).to include("#{caller.id}-shutdown")
+      end
+    end
   end
 
   describe '#on_worker_processed' do
@@ -157,6 +177,51 @@ RSpec.describe_current do
     before do
       listener.on_consumer_consume(event)
       listener.on_consumer_consumed(event)
+    end
+
+    it 'expect to remove job from running' do
+      expect(sampler.jobs).to be_empty
+    end
+  end
+
+  describe '#on_consumer_eof' do
+    before { listener.on_consumer_eof(event) }
+
+    it 'expect not to increase batches count' do
+      expect(sampler.counters[:batches]).to eq(0)
+    end
+
+    it 'expect not to increase messages count' do
+      expect(sampler.counters[:messages]).to eq(0)
+    end
+
+    it 'expect to register the job execution' do
+      expect(sampler.jobs).not_to be_empty
+    end
+
+    it 'expect to have job details' do
+      job = sampler.jobs.values.first
+
+      expect(job.keys).to include(:updated_at)
+      expect(job[:topic]).to eq('test')
+      expect(job[:partition]).to eq(0)
+      expect(job[:first_offset]).to eq(0)
+      expect(job[:last_offset]).to eq(1)
+      expect(job[:processing_lag]).to eq(1_000)
+      expect(job[:consumption_lag]).to eq(0)
+      expect(job[:committed_offset]).to eq(0)
+      expect(job[:messages]).to eq(1)
+      expect(job[:consumer]).to eq(caller.class.to_s)
+      expect(job[:consumer_group]).to eq(caller.topic.consumer_group.id)
+      expect(job[:type]).to eq('eofed')
+      expect(job[:tags]).to eq(caller.tags)
+    end
+  end
+
+  describe '#on_consumer_eofed' do
+    before do
+      listener.on_consumer_eof(event)
+      listener.on_consumer_eofed(event)
     end
 
     it 'expect to remove job from running' do