+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - **[Feature]** Support per request policies for inspection and operations limitation.
 - **[Feature]** Provide Search capabilities in the Explorer (Pro).
 - **[Feature]** Provide dark mode.
+- [Enhancement] Support Karafka `#eofed` consumer action.
 - [Enhancement] Provide topics watermarks inspection page (Pro).
 - [Enhancement] Use Turbo to improve usability.
 - [Enhancement] Round poll age reporting to precision of 2 reducing the payload size.